Feed int8 SDF directly to VTK contour, eliminate float32 volume copy
VTK's contour filter handles int8 scalars natively. The .astype(np.float32)
was unnecessary and allocated 4x the SDF size. At 504^3 this saves ~384 MB.
The full Cython fast path is now all-integer: CDT int8 -> spatial conv int8 ->
VTK contour on int8. Float only enters at MC vertex interpolation (inside VTK).

+++ src/voxelcad/voxel_model.py
@@ -370,7 +370,7 @@ class VoxelModel:
         ugrid = UniformGrid()
         ugrid.dimensions = rv  # point grid matches SDF array shape
         ugrid.spacing = vsv
-        ugrid.point_data['dist'] = dist.ravel(order='F').astype(np.float32)
+        ugrid.point_data['dist'] = dist.ravel(order='F')
         del dist
         # Marching cubes via contour on the signed distance field
         _t0 = time.time()
